+++ packages/react-native-reanimated/Common/cpp/reanimated/CSS/configs/common.h
@@ -1,38 +1,16 @@
 #pragma once
 
 #include <jsi/jsi.h>
-#include <reanimated/CSS/easing/EasingFunctions.h>
+#include <reanimated/CSS/easing/EasingConfigs.h>
 
 #include <string>
 #include <utility>
-#include <variant>
-#include <vector>
 
 namespace reanimated::css {
 
-struct LinearEasing {};
-
-struct CubicBezierEasing {
-  double x1, y1, x2, y2;
-};
-
-struct StepsEasing {
-  std::vector<double> pointsX;
-  std::vector<double> pointsY;
-};
-
-struct LinearStopsEasing {
-  std::vector<double> pointsX;
-  std::vector<double> pointsY;
-};
-
-using EasingConfig = std::variant<LinearEasing, CubicBezierEasing, StepsEasing, LinearStopsEasing>;
-
 double getDuration(jsi::Runtime &rt, const jsi::Object &config);
 
-EasingConfig getEasingConfig(jsi::Runtime &rt, const jsi::Value &easingConfig);
 EasingConfig getEasingConfig(jsi::Runtime &rt, const jsi::Object &config);
-EasingFunction getEasingFunctionFromConfig(const EasingConfig &easingConfig);
 
 double getDelay(jsi::Runtime &rt, const jsi::Object &config);
